requirements.txt
datetime
dateutil
google-auth
google-cloud-bigquery
google-cloud-secret-manager
streamlit
streamlit_chat
streamlit_option_menu
streamlit_authenticator
pandas
openai
altair
db-dtypes
plotly
aiohttp
async-timeout
dataclasses-json
tiktoken
langsmith
numpy
pydantic
PyYAML
requests
SQLAlchemy
tenacity
langchain-text-splitters
langchain
jsonpatch
langchain-community
langchain-experimental
langchain_openai
langchain_core
langchain-text-splitters




